+++ src/index.ts
@@ -2,7 +2,7 @@
 import $ from "jquery";
 import { vec3 } from "gl-matrix";
 import WebGL2 from "./gl";
-import { Teapot } from "./renderables/teapot";
+import { Teapot } from "./models/teapot";
 import { CrepuscularRay } from "./shaders/crepuscular_ray/shader";
 import { TransparentShader } from "./shaders/transparent/shader";
 import { FXAA } from "./shaders/fxaa/shader";
