+++ src/light.ts
@@ -33,14 +33,21 @@ export class Light {
     this.color = props.color ?? vec3.fromValues(1, 1, 1);
     this.power = props.power ?? 1;
     this.radius = props.radius ?? 1;
-    this.matrix = new Matrix(gl, mat4.fromTranslation(mat4.create(), this.position));
+    this.matrix = new Matrix(
+      gl,
+      mat4.fromTranslation(mat4.create(), this.position)
+    );
 
     const degrees = props.degrees ?? 20;
     this.vertices = this.getVertices(degrees);
 
     this.verticesBuffer = gl.createBuffer();
     gl.bindBuffer(this.gl.ARRAY_BUFFER, this.verticesBuffer);
-    gl.bufferData(this.gl.ARRAY_BUFFER, new Float32Array(this.vertices), gl.STATIC_DRAW);
+    gl.bufferData(
+      this.gl.ARRAY_BUFFER,
+      new Float32Array(this.vertices),
+      gl.STATIC_DRAW
+    );
   }
 
   // Points on a circle are given by the equation (x, y) = (rsin(θ), rcos(θ)).
@@ -50,7 +57,7 @@ export class Light {
     for (let d = 0; d <= 360; d += degrees) {
       vertices.push(
         this.radius * Math.sin(glMatrix.toRadian(d)),
-        this.radius * Math.cos(glMatrix.toRadian(d)),
+        this.radius * Math.cos(glMatrix.toRadian(d))
       );
     }
